채팅 메시지 위치 및 endpoint 변경 (#70)
* feat : 방 구독자들에게 메세지 전체 전달하기

https://github.com/KHUCapston-concoder/Backend/issues/58

* fix : 확인 로그 추가

https://github.com/KHUCapston-concoder/Backend/issues/58

* feat : 채팅 컨틀롤러 확인 로그 추가

https://github.com/KHUCapston-concoder/Backend/issues/58

* 자료형 변경

* fix : 최대 연결시간 변경

https://github.com/KHUCapston-concoder/Backend/issues/58

* fix : jsonObject로 변경

https://github.com/KHUCapston-concoder/Backend/issues/58

* fix : 위치 및 endpoint 변경

https://github.com/KHUCapston-concoder/Backend/issues/58
