+++ src/main/java/oncoding/concoder/controller/ChattingController.java
@@ -22,14 +22,15 @@ public class ChattingController {
   private final ChattingService chatService;
 
 
-  @MessageMapping("/rooms/chat/{roomId}")
-  public void chat(@DestinationVariable final String roomId, JSONObject ob) {
-    log.info("/rooms/chat/"+roomId+" userId:  "+ob.get("userId"));
-    log.info("/rooms/chat/"+roomId+" content: "+ob.get("content"));
-    MessageRequest request = new MessageRequest(UUID.fromString((String) ob.get("userId")), (String)ob.get("content"));
-    template.convertAndSend("/sub/rooms/chat/"+ roomId , chatService.sendMessage(request));
-    log.info("after chatting convert and send");
-  }
+//  @MessageMapping("/rooms/chat/{roomId}")
+//  public void chat(@DestinationVariable final String roomId, JSONObject ob) {
+//    log.info("/rooms/chat/"+roomId+" userId:  "+ob.get("userId"));
+//    log.info("/rooms/chat/"+roomId+" content: "+ob.get("content"));
+//    MessageRequest request = new MessageRequest(UUID.fromString((String) ob.get("userId")), (String)ob.get("content"));
+//    template.convertAndSend("/sub/rooms/chat/"+ roomId , chatService.sendMessage(request));
+//    log.info("after chatting convert and send");
+//  }
+
